Battleship Game in Python
An implementation of the battleship Game in Python, built off of a simpler version made for class. The game is structured using a class, and features functions for setting up the game board, checking for hits or misses, determining game over conditions and playing the game. A set of unit tests have been added to ensure functionality of game logic.